+++ firmware/src/drivers/FingerprintDriver.h
@@ -14,16 +14,16 @@ class FingerprintDriver {
   uint8_t getImage();
   uint8_t image2Tz(uint8_t slot);
   uint8_t createModel();
-  uint8_t storeModel(uint8_t id);
-  uint8_t loadModel(uint8_t id);
-  uint8_t deleteModel(uint8_t id);
+  uint16_t storeModel(uint16_t id);
+  uint16_t loadModel(uint16_t id);
+  uint16_t deleteModel(uint16_t id);
 
   /** Returns the first slot id in [1, maxId] not used by a stored template, or 0 if none. */
-  uint8_t findFirstFreeSlot(uint8_t maxId);
+  uint16_t findFirstFreeSlot(uint16_t maxId);
 
-  String getTemplateAsHex(uint8_t id);
+  String getTemplateAsHex(uint16_t id);
 
-  bool setTemplateFromHex(uint8_t id, const String& hexData);
+  bool setTemplateFromHex(uint16_t id, const String& hexData);
 
  private:
   HardwareSerial& serialPort_;
