+++ firmware/src/services/SyncMappingService.cpp
@@ -10,11 +10,11 @@ SyncMappingService::SyncMappingService(drivers::FingerprintDriver& fingerprint,
   }
 }
 
-bool SyncMappingService::enqueue(uint8_t employeeId, uint8_t fingerprintSlot) {
+bool SyncMappingService::enqueue(uint32_t employeeId, uint16_t fingerprintSlot) {
   for (uint8_t i = 0; i < kMaxPending; i++) {
     if (!pending_[i].inUse) {
       pending_[i] = {true, employeeId, fingerprintSlot, 0, millis()};
-      Serial.printf("[SyncMapping] Enqueued employee=%u slot=%u\n",
+      Serial.printf("[SyncMapping] Enqueued employee=%lu slot=%u\n",
                     employeeId, fingerprintSlot);
       return true;
     }
@@ -41,21 +41,21 @@ void SyncMappingService::tick(const models::DeviceConfig& config,
     }
 
     entry.attempts++;
-    Serial.printf("[SyncMapping] Attempt %u/%u for employee=%u slot=%u\n",
+    Serial.printf("[SyncMapping] Attempt %u/%u for employee=%lu slot=%u\n",
                   entry.attempts, maxAttempts, entry.employeeId, entry.fingerprintSlot);
 
     const bool ok = network_.registerSyncMappingCallback(
         config, apiKey, entry.employeeId, entry.fingerprintSlot);
 
     if (ok) {
-      Serial.printf("[SyncMapping] Success for employee=%u slot=%u\n",
+      Serial.printf("[SyncMapping] Success for employee=%lu slot=%u\n",
                     entry.employeeId, entry.fingerprintSlot);
       entry.inUse = false;
       return;
     }
 
     if (entry.attempts >= maxAttempts) {
-      Serial.printf("[SyncMapping] Exhausted retries for employee=%u slot=%u — rolling back sensor.\n",
+      Serial.printf("[SyncMapping] Exhausted retries for employee=%lu slot=%u — rolling back sensor.\n",
                     entry.employeeId, entry.fingerprintSlot);
       const uint8_t delResult = fingerprint_.deleteModel(entry.fingerprintSlot);
       Serial.printf("[SyncMapping] deleteModel(%u) -> %u\n",
